refactor: unify the two colliding user query key factories
`users/queries.ts` and `administration/queries.ts` each exported a
`userQueryKeys` factory with a different hierarchy and a different id
type. The SSE handler bound administration's factory to the `users`
domain, while the infinite user list was keyed by the other one.

Invalidation reached across that boundary only by prefix accident:
administration's `lists()` (`["users","list"]`) happened to prefix-match
`["users","list","infinite",...]` from the users factory. Aligning the two
onto administration's `["users","infiniteList"]` shape would have silently
stopped the infinite list refreshing, with no type error and no failing
test.

Collapse to one factory in `users/queries.ts`, on the hierarchy the rest
of the codebase already uses (indexes, groups, uploads all nest infinite
lists under `list`), with administration's `number` id — which is what the
SSE frame and `getUser` actually carry. The dead `detail(user_id: string)`
variant had no callers.

Move the user hooks to `users/queries.ts` alongside it, so the feature
owns its own server state and its components stop importing them from
administration. `administration/queries.ts` keeps settings and the
administrator-roles list.

Give `useListUsers` its own `nested()` key. It was keyed `list([])`, which
collapses to `["users","list"]` — byte-identical to `lists()` — parking a
live cache entry on the invalidation prefix.

Add tests that hold the invariant explicitly: the SSE handler test now
seeds a real QueryClient and asserts every user list variant is marked
stale, and a factory test asserts the hierarchy nests as intended.
