+++ apps/web/src/users/components/Password.tsx
@@ -1,4 +1,3 @@
-import { useUpdateUser } from "@administration/queries";
 import BoxGroup from "@base/BoxGroup";
 import BoxGroupHeader from "@base/BoxGroupHeader";
 import BoxGroupSection from "@base/BoxGroupSection";
@@ -9,6 +8,7 @@ import InputGroup from "@base/InputGroup";
 import InputSimple from "@base/InputSimple";
 import RelativeTime from "@base/RelativeTime";
 import SaveButton from "@base/SaveButton";
+import { useUpdateUser } from "@users/queries";
 import { useForm } from "react-hook-form";
 
 type PasswordProps = {
