+++ apps/web/src/users/components/UserGroups.tsx
@@ -1,4 +1,3 @@
-import { useUpdateUser } from "@administration/queries";
 import { useFuse } from "@app/fuse";
 import BoxGroup from "@base/BoxGroup";
 import BoxGroupSection from "@base/BoxGroupSection";
@@ -10,6 +9,7 @@ import QueryError from "@base/QueryError";
 import { RadioGroup, RadioGroupItem } from "@base/RadioGroup";
 import { useListGroups } from "@groups/queries";
 import type { GroupMinimal } from "@groups/types";
+import { useUpdateUser } from "@users/queries";
 import { X } from "lucide-react";
 
 /** A stable empty fallback so `useFuse` doesn't reset its term while loading */
