+++ apps/web/src/users/components/UsersList.tsx
@@ -1,10 +1,10 @@
-import { useFindUsers } from "@administration/queries";
 import Box from "@base/Box";
 import BoxGroup from "@base/BoxGroup";
 import { Empty, EmptyDescription, EmptyMedia, EmptyTitle } from "@base/Empty";
 import LoadingPlaceholder from "@base/LoadingPlaceholder";
 import Pagination from "@base/Pagination";
 import QueryError from "@base/QueryError";
+import { useFindUsers } from "@users/queries";
 import { Users } from "lucide-react";
 import type { User } from "../types";
 import { UserItem } from "./UserItem";
